requirements.txt
# Isi api_ingestion/requirements.txt
Flask
psycopg2-binary
flask-cors
gunicorn